rgetattr-setattr
